coding-assistant/requirements.txt
requests
transformers
scikit-learn
numpy
torch
sentence-transformers
pyyaml
python-box
python-dotenv
optimum
together
pytest